refactor(meai-10-5): T-021 DR-11 migration sweep — delete old single-arity types
Closes the DR-11 contract break by deleting the old single-arity
IAgentStep<TState> interface, the abstract AgentStepBase<TState> class,
and the AgentStepBaseTests.TestAgentStep fixture; renames the new
two-arity files (IAgentStepT2.cs, AgentStepBaseT2.cs) to canonical
IAgentStep.cs / AgentStepBase.cs. No [Obsolete] shim — clean break,
verified zero production consumers via pre-flight grep.

One in-kind touch: src/Strategos.Agents.Mcp.Tests/PackageMetadataTests.cs
referenced typeof(IAgentStep<>) purely to grab the core assembly handle;
swapped to typeof(IAgentStep<,>). Not a real consumer of the old contract.

Reflection invariant tests (Unit/MigrationInvariantTests.cs) guard
against either single-arity type — or the TestAgentStep fixture —
reappearing in the production or test assemblies.

Implements: DR-1, DR-11.
