+++ src/Strategos.Agents/AgentStepBase.cs
@@ -4,80 +4,37 @@
 // </copyright>
 // =============================================================================
 
+using Microsoft.Extensions.AI;
 using Strategos.Abstractions;
 using Strategos.Agents.Abstractions;
-using Strategos.Agents.Models;
+using Strategos.Agents.Configuration;
+using Strategos.Agents.Exceptions;
 using Strategos.Steps;
-
-using Microsoft.Extensions.AI;
 
 namespace Strategos.Agents;
 
 /// <summary>
-/// Base class for LLM-powered workflow steps with context assembly support.
+/// Sealed orchestrator for LLM-powered workflow steps with typed structured results (DR-1).
+/// Constructed via <c>AgentStepBuilder&lt;TState, TResult&gt;</c>; never subclassed.
 /// </summary>
-/// <typeparam name="TState">The type of workflow state.</typeparam>
-/// <remarks>
-/// <para>
-/// AgentStepBase provides a structured approach to building LLM-powered steps:
-/// <list type="bullet">
-///   <item><description>Context assembly from multiple sources (state, RAG, literals)</description></item>
-///   <item><description>Message building with system prompt, context, and user prompt</description></item>
-///   <item><description>Response handling and state updates</description></item>
-/// </list>
-/// </para>
-/// <para>
-/// Subclasses must implement:
-/// <list type="bullet">
-///   <item><description><see cref="GetSystemPrompt"/> - Define agent behavior</description></item>
-///   <item><description><see cref="GetUserPrompt"/> - Build the user query from state</description></item>
-///   <item><description><see cref="ApplyResultAsync"/> - Update state with LLM response</description></item>
-/// </list>
-/// </para>
-/// </remarks>
-public abstract class AgentStepBase<TState> : IAgentStep<TState>
+/// <typeparam name="TState">Workflow state type.</typeparam>
+/// <typeparam name="TResult">Typed structured result produced by the agent's chat client.</typeparam>
+public sealed class AgentStepBase<TState, TResult> : IAgentStep<TState, TResult>
     where TState : class, IWorkflowState
 {
-    private readonly IChatClient _chatClient;
-    private readonly IContextAssembler<TState>? _contextAssembler;
+    /// <summary>Default upper bound on tool-call iterations per request (DR-8).</summary>
+    public const int DefaultMaxToolIterations = 8;
 
-    /// <summary>
-    /// Initializes a new instance of the <see cref="AgentStepBase{TState}"/> class.
-    /// </summary>
-    /// <param name="chatClient">The chat client for LLM interactions.</param>
-    /// <param name="contextAssembler">Optional context assembler for RAG support.</param>
-    protected AgentStepBase(
+    private readonly IChatClient _chatClient;
+    private readonly AgentStepConfiguration<TState, TResult> _configuration;
+
+    internal AgentStepBase(
         IChatClient chatClient,
-        IContextAssembler<TState>? contextAssembler = null)
+        AgentStepConfiguration<TState, TResult> configuration)
     {
         _chatClient = chatClient ?? throw new ArgumentNullException(nameof(chatClient));
-        _contextAssembler = contextAssembler;
+        _configuration = configuration ?? throw new ArgumentNullException(nameof(configuration));
     }
-
-    /// <inheritdoc/>
-    public abstract string GetSystemPrompt();
-
-    /// <inheritdoc/>
-    public abstract Type? GetOutputSchemaType();
-
-    /// <summary>
-    /// Gets the user prompt based on the current workflow state.
-    /// </summary>
-    /// <param name="state">The current workflow state.</param>
-    /// <returns>The user prompt string.</returns>
-    protected abstract string GetUserPrompt(TState state);
-
-    /// <summary>
-    /// Applies the LLM response to update the workflow state.
-    /// </summary>
-    /// <param name="state">The current workflow state.</param>
-    /// <param name="response">The LLM response text.</param>
-    /// <param name="cancellationToken">Cancellation token.</param>
-    /// <returns>The step result with updated state.</returns>
-    protected abstract Task<StepResult<TState>> ApplyResultAsync(
-        TState state,
-        string response,
-        CancellationToken cancellationToken);
 
     /// <inheritdoc/>
     public async Task<StepResult<TState>> ExecuteAsync(
@@ -85,52 +42,88 @@ public abstract class AgentStepBase<TState> : IAgentStep<TState>
         StepContext context,
         CancellationToken cancellationToken)
     {
-        // Assemble context if assembler is available
-        var assembledContext = _contextAssembler != null
-            ? await _contextAssembler.AssembleAsync(state, context, cancellationToken).ConfigureAwait(false)
-            : AssembledContext.Empty;
+        var messages = BuildMessages(state);
 
-        // Build messages for the LLM
-        var messages = BuildMessages(state, assembledContext);
+        ChatResponse<TResult>? response;
+        try
+        {
+            response = await _chatClient.GetResponseAsync<TResult>(
+                    messages,
+                    _configuration.ChatOptions,
+                    cancellationToken: cancellationToken)
+                .ConfigureAwait(false);
+        }
+        catch (OperationCanceledException)
+        {
+            // Cancellation is not a domain failure (DR-10) — propagate unwrapped.
+            throw;
+        }
+        catch (ArgumentNullException)
+        {
+            // MEAI's typed extension throws ArgumentNullException when the underlying
+            // chat client returns a null ChatResponse. That's the DR-10 boundary we
+            // own: re-classify as AGAG006 with no partial state mutation.
+            throw new AgentChatResponseException(
+                "Chat client returned a null response.");
+        }
 
-        // Get response from LLM
-        var response = await _chatClient.GetResponseAsync(
-            messages,
-            options: null,
-            cancellationToken).ConfigureAwait(false);
+        if (response is null)
+        {
+            // Defensive: should not happen via MEAI typed extension (it throws ANE),
+            // but if a direct caller produces null we surface the same DR-10 contract.
+            throw new AgentChatResponseException(
+                "Chat client returned a null response.");
+        }
 
-        var responseText = response.Text ?? string.Empty;
+        // DR-8: detect the FunctionInvokingChatClient iteration cap. MEAI 10.5 does not
+        // throw when it reaches MaximumIterationsPerRequest — it logs
+        // "Reached maximum iteration count of {N}. Stopping function invocation loop."
+        // and returns the latest inner response as-is. That response retains
+        // FinishReason == ToolCalls (the model still wanted to call tools when the
+        // middleware bailed). We surface that condition as AGAG005 with the captured
+        // PartialTrace. This check MUST precede the AGAG006 empty-payload check,
+        // because the capped response will also have an empty Text / no TryGetResult.
+        if (response.FinishReason == ChatFinishReason.ToolCalls)
+        {
+            var maxIterations = _configuration.MaxToolIterations ?? DefaultMaxToolIterations;
+            // Snapshot the messages into an immutable IReadOnlyList. We do NOT alias
+            // response.Messages directly — callers must observe a stable trace even
+            // if the response is later mutated by another consumer.
+            var partialTrace = response.Messages.ToArray();
+            throw new AgentToolLoopException(maxIterations, partialTrace);
+        }
 
-        // Apply result and return updated state
-        return await ApplyResultAsync(state, responseText, cancellationToken).ConfigureAwait(false);
+        var hasResult = response.TryGetResult(out var typedResult) && typedResult is not null;
+        if (!hasResult && string.IsNullOrEmpty(response.Text))
+        {
+            // DR-10: empty payload (no Result, no Text) — distinct from a malformed
+            // structured-output failure (AGAG002). State is untouched (apply-result
+            // never runs); the caller's TState instance remains reference-equal.
+            throw new AgentChatResponseException(
+                "Chat client returned an empty response (no text and no structured result).");
+        }
+
+        if (!hasResult)
+        {
+            // DR-3 / DR-10 no-silent-fallback: throw with AGAG002 and the raw payload.
+            // The exception ctor handles ≤4 KB truncation; apply-result hook is NOT invoked.
+            throw new AgentStructuredOutputException(response.Text);
+        }
+
+        return await _configuration.ApplyResult(state, typedResult!, cancellationToken).ConfigureAwait(false);
     }
 
     /// <summary>
-    /// Builds the message list for the LLM call.
+    /// Builds the message sequence sent to the chat client for a given state.
     /// </summary>
     /// <param name="state">The current workflow state.</param>
-    /// <param name="context">The assembled context.</param>
-    /// <returns>The list of chat messages.</returns>
-    /// <remarks>
-    /// This method is internal for testing purposes.
-    /// </remarks>
-    internal IList<ChatMessage> BuildMessages(TState state, AssembledContext context)
+    /// <returns>The ordered system + user messages.</returns>
+    internal IList<ChatMessage> BuildMessages(TState state)
     {
-        var messages = new List<ChatMessage>();
-
-        // Add system prompt
-        messages.Add(new ChatMessage(ChatRole.System, GetSystemPrompt()));
-
-        // Add context if non-empty
-        if (!context.IsEmpty)
+        return new List<ChatMessage>
         {
-            var contextText = context.ToPromptString();
-            messages.Add(new ChatMessage(ChatRole.User, $"Context:\n{contextText}"));
-        }
-
-        // Add user prompt
-        messages.Add(new ChatMessage(ChatRole.User, GetUserPrompt(state)));
-
-        return messages;
+            new(ChatRole.System, _configuration.SystemPrompt(state)),
+            new(ChatRole.User, _configuration.UserPrompt(state)),
+        };
     }
 }
